Added "pipeReplacer" and "prefixCellValues" options to table entries. (#46)
pipeReplacer: This option supports use cases that require a different mechanism for
escaping pipes within tables. For example, including an Obsidian link with a display
name requires escaping with a slash (i.e. "\|") instead of using an HTML entity.

prefixCellValues: The default table rendering is to include the "prefix" string before
the value in every table cell, in addition to prefixing each individual table row. This
option allows this behavior to be disabled, so that cell values are not prefixed. (It
is possible the default behavior is a bug, but by providing a flag to change it this PR
maintains backwards compatibility.)